Added ability to user gradients on bullet charts.
While using bullet charts, you now have the extra option "gradient"
which can be set to true. If it's set to true, a gradient spectrum is
created for the rangeColors that are passed. If only 1 color is passed,
all ranges are that color. If more than 2 colors are passed, the
gradient shifts between each color. This has been implemented with
Rainbowvis-js (https://github.com/anomal/RainbowVis-JS). It has been
included and linked in the makefile.