Fix tooltip width and position when content change
When the tooltip is on the border right for sample, and the mouse pointer is moved over another region implying another wider content, a scollbar may appear. This is due to the computation of tooltips before the content update.

This PR ensures the size is computed after the content is updated.